counting class occurences and summarization over countries per questin and class
